Refactor to use latest llvm/lldb
